+++ NEWS.md
@@ -1,6 +1,10 @@
+## tsfgrnn 0.3.0
+
+* Automatic computation of smoothing parameter can be done with rolling origin or fixed origin (less accurate, but faster).
+
 ## tsfgrnn 0.2.0
 
-* Uses Rcpp for speeding up computation of predictions
+* Uses Rcpp for speeding up computation of forecasts.
 
 
 ## tsfgrnn 0.1.0